sdn-obfuscation-network/requirements.txt
etcd3==0.12.0
grpcio==1.70.0
networkx==3.1
psutil==7.2.2
protobuf==3.20.3
six==1.17.0
tenacity==9.0.0
